@@ -17,63 +17,63 @@
     <parent>
         <groupId>org.springframework.boot</groupId>
         <artifactId>spring-boot-starter-parent</artifactId>
-        <version>3.0.5</version>
+        <version>3.2.2</version>
         <relativePath/>
     </parent>
 
     <properties>
-        <!--
-        <java.version>17</java.version>
+        <!---->
+        <java.version>21</java.version>
         <maven.compiler.source>${java.version}</maven.compiler.source>
         <maven.compiler.target>${java.version}</maven.compiler.target>
         <maven.compiler.compilerVersion>${java.version}</maven.compiler.compilerVersion>
-        -->
+
 
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <project.reporting.outputEncoding>${project.build.sourceEncoding}</project.reporting.outputEncoding>
         <maven.build.timestamp.format>yyyy-MM-dd HH:mm:ss</maven.build.timestamp.format>
 
-        <clever-framework.version>1.1.0-SNAPSHOT</clever-framework.version>
+        <clever-framework.version>1.2.0-SNAPSHOT</clever-framework.version>
         <!-- CI-friendly property 必须使用这个名称，否则IDEA报错-->
         <revision>${clever-framework.version}</revision>
 
-        <!-- spring jar -->
-        <spring-boot.version>3.0.3</spring-boot.version>
+
+        <spring-boot-admin.version>3.2.2</spring-boot-admin.version>
 
 
-        <spring-boot-admin.version>3.0.2</spring-boot-admin.version>
+        <guava.version>33.0.0-jre</guava.version>
+        <jsqlparse.version>4.7</jsqlparse.version>
 
-        <guava.version>31.1-jre</guava.version>
-        <jsqlparse.version>4.5</jsqlparse.version>
-
-        <mapstruct.version>1.5.3.Final</mapstruct.version>
+        <mapstruct.version>1.5.5.Final</mapstruct.version>
 
         <!-- apache jar -->
-        <commons-fileupload.version>1.4</commons-fileupload.version>
+        <commons-fileupload.version>1.5</commons-fileupload.version>
         <commons-collections4.version>4.4</commons-collections4.version>
-        <commons-io.version>2.11.0</commons-io.version>
+        <commons-io.version>2.15.1</commons-io.version>
         <commons-beanutils.version>1.9.4</commons-beanutils.version>
 
         <validation-api.version>2.0.1.Final</validation-api.version>
 
-        <jjwt.version>0.9.1</jjwt.version>
+        <jjwt.version>0.12.3</jjwt.version>
         <datanucleus.version>5.2.1</datanucleus.version>
 
+
         <springdoc.version>1.7.0</springdoc.version>
-        <swagger.version>2.2.2</swagger.version>
+        <swagger.version>2.2.20</swagger.version>
 
-        <minio.version>8.5.2</minio.version>
-        <caffeine.version>3.1.5</caffeine.version>
 
-        <hutool.version>5.8.16</hutool.version>
+        <minio.version>8.5.8</minio.version>
+        <caffeine.version>3.1.8</caffeine.version>
 
-        <!--
-        <mybatis.version>3.5.9</mybatis.version>
-        <mybatis-spring.version>2.0.7</mybatis-spring.version>
-        <mybatis-spring-boot-starter.version>2.5.2</mybatis-spring-boot-starter.version>
-        -->
-        <mybatis-plus.version>3.5.3.1</mybatis-plus.version>
-        <mybatis-plus-dynamic-datasource.version>3.6.1</mybatis-plus-dynamic-datasource.version>
+        <hutool.version>5.8.26</hutool.version>
+
+        <!-- -->
+        <mybatis.version>3.5.15</mybatis.version>
+        <mybatis-spring.version>3.0.3</mybatis-spring.version>
+        <mybatis-spring-boot-starter.version>3.0.3</mybatis-spring-boot-starter.version>
+
+        <mybatis-plus.version>3.5.5</mybatis-plus.version>
+        <mybatis-plus-dynamic-datasource.version>4.2.0</mybatis-plus-dynamic-datasource.version>
 
     </properties>
 
