@@ -7,7 +7,7 @@ import java.security.SecureRandom;
 /**
  *
  */
-public final class AppNanoIdUtils {
+public final class NanoIdUtils {
     public static final SecureRandom DEFAULT_NUMBER_GENERATOR = new SecureRandom();
     // 随机生成的字符内容
     public static final char[] DEFAULT_ALPHABET = "23456789abcdefghijklmnopqrstuvwxyzABCDEFGHJKLMNPQRSTUVWXYZ".toCharArray();
@@ -15,7 +15,7 @@ public final class AppNanoIdUtils {
     // 随机生成 ID 的位数
     public static final int DEFAULT_SIZE = 16;
 
-    private AppNanoIdUtils() {
+    private NanoIdUtils() {
     }
 
     public static String randomNanoId() {
