@@ -8,7 +8,7 @@ import java.util.List;
  * @author toquery
  * @version 1
  */
-public interface AppEntityTree<E> extends Comparable<E> {
+public interface EntityTree<E> extends Comparable<E> {
 
     Long getId();
 
