@@ -3,7 +3,7 @@ package io.github.toquery.framework.core.entity;
 /**
  * 是否是假删除 逻辑
  */
-public interface AppEntityLogicDel {
+public interface EntityLogicDel {
 
     Boolean getDeleted() ;
 
