@@ -1,7 +1,7 @@
 package io.github.toquery.framework.crud.controller;
 
 import io.github.toquery.framework.crud.service.impl.AppBFFServiceImpl;
-import io.github.toquery.framework.core.entity.AppBaseEntity;
+import io.github.toquery.framework.core.entity.BaseEntity;
 import io.github.toquery.framework.webmvc.controller.AppBaseWebMvcController;
 
 /**
@@ -9,7 +9,7 @@ import io.github.toquery.framework.webmvc.controller.AppBaseWebMvcController;
  * @param <BS> bff service
  * @param <BD> bff dao
  */
-public class AppBaseBFFController<E extends AppBaseEntity, BS extends AppBFFServiceImpl<E, BD>, BD extends AppJpaBaseRepository<E>> extends AppBaseWebMvcController {
+public class AppBaseBFFController<E extends BaseEntity, BS extends AppBFFServiceImpl<E, BD>, BD extends AppJpaBaseRepository<E>> extends AppBaseWebMvcController {
     protected BS bffService;
 
     public AppBaseBFFController(BS bffService) {
