@@ -1,6 +1,6 @@
 package io.github.toquery.framework.jpa.audit;
 
-import io.github.toquery.framework.core.entity.AppBaseEntity;
+import io.github.toquery.framework.core.entity.BaseEntity;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.context.annotation.Scope;
 import org.springframework.data.domain.AfterDomainEventPublication;
@@ -54,19 +54,19 @@ public class AppEntityD3Listener {
     }
 
     @PostLoad
-    public void onPostLoad(AppBaseEntity appBaseEntity) {
+    public void onPostLoad(BaseEntity appBaseEntity) {
         log.debug("Spring DDD Model: AppEntityD3Listener -- {} -- @PostLoad", appBaseEntity.getClass().getSimpleName());
         appAuditorHandles.stream().filter(AppAuditorHandler::enable).forEach(item -> item.onPostLoad(appBaseEntity));
     }
 
     @PrePersist
-    public void onPrePersist(AppBaseEntity appBaseEntity) {
+    public void onPrePersist(BaseEntity appBaseEntity) {
         log.debug("Spring DDD Model: AppEntityD3Listener ---- @PrePersist");
         appAuditorHandles.stream().filter(AppAuditorHandler::enable).forEach(item -> item.onPrePersist(appBaseEntity));
     }
 
     @PostPersist
-    public void onPostPersist(AppBaseEntity appBaseEntity) {
+    public void onPostPersist(BaseEntity appBaseEntity) {
         log.debug("Spring DDD Model: AppEntityD3Listener ---- @PostPersist");
         appAuditorHandles.stream().filter(AppAuditorHandler::enable).forEach(item -> item.onPostPersist(appBaseEntity));
     }
@@ -76,26 +76,26 @@ public class AppEntityD3Listener {
      * todo BUG，修改数据时不经过这个
      */
     @PreUpdate
-    public void onPreUpdate(AppBaseEntity appBaseEntity) {
+    public void onPreUpdate(BaseEntity appBaseEntity) {
         log.debug("Spring DDD Model: AppEntityD3Listener - @PreUpdate");
         appAuditorHandles.stream().filter(AppAuditorHandler::enable).forEach(item -> item.onPreUpdate(appBaseEntity));
     }
 
     @PostUpdate
-    public void onPostUpdate(AppBaseEntity appBaseEntity) {
+    public void onPostUpdate(BaseEntity appBaseEntity) {
         log.debug("Spring DDD Model: AppEntityD3Listener ---- @PostUpdate");
         appAuditorHandles.stream().filter(AppAuditorHandler::enable).forEach(item -> item.onPostUpdate(appBaseEntity));
 
     }
 
     @PreRemove
-    public void onPreRemove(AppBaseEntity appBaseEntity) {
+    public void onPreRemove(BaseEntity appBaseEntity) {
         log.debug("Spring DDD Model: AppEntityD3Listener - @PreRemove");
         appAuditorHandles.stream().filter(AppAuditorHandler::enable).forEach(item -> item.onPreRemove(appBaseEntity));
     }
 
     @PostRemove
-    public void onPostRemove(AppBaseEntity appBaseEntity) {
+    public void onPostRemove(BaseEntity appBaseEntity) {
         log.debug("Spring DDD Model: AppEntityD3Listener ---- @PostRemove");
         appAuditorHandles.stream().filter(AppAuditorHandler::enable).forEach(item -> item.onPostRemove(appBaseEntity));
 
