@@ -3,15 +3,18 @@ package io.github.toquery.framework.system.entity;
 import com.baomidou.mybatisplus.annotation.TableField;
 import com.baomidou.mybatisplus.annotation.TableLogic;
 import com.baomidou.mybatisplus.annotation.TableName;
-import io.github.toquery.framework.core.domain.AppEntitySort;
-import io.github.toquery.framework.core.domain.AppEntityTree;
-import io.github.toquery.framework.core.entity.AppBaseEntity;
-import io.github.toquery.framework.core.entity.AppEntityLogicDel;
+import io.github.toquery.framework.common.constant.EntityFieldConstant;
+import io.github.toquery.framework.core.domain.EntitySort;
+import io.github.toquery.framework.core.domain.EntityTree;
+import io.github.toquery.framework.core.entity.BaseEntity;
+import io.github.toquery.framework.core.entity.EntityLogicDel;
 import lombok.Getter;
 import lombok.Setter;
 import org.hibernate.annotations.ColumnDefault;
 import org.hibernate.annotations.SQLDelete;
-import org.hibernate.annotations.Where;
+import org.hibernate.annotations.SQLRestriction;
+import org.hibernate.annotations.SQLRestriction;
+import io.github.toquery.framework.common.constant.EntityFieldConstant;
 import org.hibernate.validator.constraints.Length;
 
 import jakarta.persistence.Column;
@@ -32,9 +35,9 @@ import java.util.List;
 @Setter
 @TableName(value = "sys_area")
 @Table(name = "sys_area")
-@Where(clause = "deleted = false")
+@SQLRestriction(value = EntityFieldConstant.DOMAIN_FIELD_SOFT_DEL + " = false")
 @SQLDelete(sql ="UPDATE sys_area SET deleted = true WHERE id = ?")
-public class SysArea extends AppBaseEntity implements AppEntityTree<SysArea>, AppEntitySort, AppEntityLogicDel {
+public class SysArea extends BaseEntity implements EntityTree<SysArea>, EntitySort, EntityLogicDel {
 
     /**
      * 名称
