@@ -1,16 +1,17 @@
 package io.github.toquery.framework.system.entity;
 
-import io.github.toquery.framework.core.domain.AppEntitySort;
+import io.github.toquery.framework.core.domain.EntitySort;
 import io.github.toquery.framework.core.log.annotation.AppLogField;
-import io.github.toquery.framework.core.entity.AppBaseEntity;
-import io.github.toquery.framework.core.entity.AppEntityLogicDel;
+import io.github.toquery.framework.core.entity.BaseEntity;
+import io.github.toquery.framework.core.entity.EntityLogicDel;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 import org.hibernate.annotations.ColumnDefault;
 import org.hibernate.annotations.SQLDelete;
-import org.hibernate.annotations.Where;
+import org.hibernate.annotations.SQLRestriction;
+import io.github.toquery.framework.common.constant.EntityFieldConstant;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -28,9 +29,9 @@ import java.util.List;
 @NoArgsConstructor
 @AllArgsConstructor
 @Table(name = "sys_dict")
-@Where(clause = "deleted = false")
+@SQLRestriction(value = EntityFieldConstant.DOMAIN_FIELD_SOFT_DEL + " = false")
 @SQLDelete(sql ="UPDATE sys_dict SET deleted = true WHERE id = ?")
-public class SysDict extends AppBaseEntity implements AppEntitySort, AppEntityLogicDel {
+public class SysDict extends BaseEntity implements EntitySort, EntityLogicDel {
 
     /**
      * 字典名称
